Move test exec env tests to separate module
