+++ buildpacks/dotnet/tests/mod.rs
@@ -6,6 +6,7 @@ mod dotnet_publish_test;
 mod dotnet_restore_tools_test;
 mod nuget_layer_test;
 mod sdk_installation_test;
+mod test_execution_environment_test;
 
 pub(crate) fn default_build_config(fixture_path: impl AsRef<Path>) -> BuildConfig {
     #[cfg(target_arch = "x86_64")]
